backend/pyproject.toml
[project]
name = "runcrew-backend"
version = "0.1.0"
description = "RunCrew 런닝 코스 공유 앱 백엔드"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "geoalchemy2>=0.15.2",
    # Auth / Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.2.1",
    # HTTP client (social login API calls)
    "httpx>=0.28.1",
    # Settings
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    # File handling
    "aiofiles>=24.1.0",
    "Pillow>=11.1.0",
    # Utilities
    "python-dateutil>=2.9.0",
    # Geo
    "shapely>=2.0.0",
    # DI
    "dependency-injector>=4.42.0",
    # File import (GPX/FIT parsing)
    "gpxpy>=1.6.0",
    "fitparse>=0.0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
